+++ app-android/src/main/java/io/github/droidkaigi/confsched/App.kt
@@ -15,7 +15,6 @@ class App : Application(), ServerEnvironmentModule.HasServerEnvironment {
 
     override fun onCreate() {
         super.onCreate()
-        AppContext.setUp(applicationContext)
         AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
     }
 }
