+++ app-android/src/main/java/io/github/droidkaigi/confsched/share/SaveToDisk.kt
@@ -1,29 +1,29 @@
-package io.github.droidkaigi.confsched.ui
+package io.github.droidkaigi.confsched.share
 
-import android.graphics.Bitmap
+import android.content.Context
+import android.graphics.Bitmap.CompressFormat.PNG
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.asAndroidBitmap
-import io.github.droidkaigi.confsched.AppContext
-import kotlinx.datetime.Clock
+import kotlinx.datetime.Clock.System
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
 import java.io.File
 import java.io.FileOutputStream
 
-actual fun ImageBitmap.saveToDisk(): String {
-    val timestamp = Clock.System.now()
+fun ImageBitmap.saveToDisk(context: Context): String {
+    val timestamp = System.now()
         .toLocalDateTime(TimeZone.UTC)
         .toString()
         .replace(":", "")
         .replace(".", "")
     val fileName = "shared_image_$timestamp.png"
 
-    val cachePath = File(AppContext.get().cacheDir, "images")
+    val cachePath = File(context.cacheDir, "images")
     cachePath.mkdirs()
     val file = File(cachePath, fileName)
     val outputStream = FileOutputStream(file)
 
-    this.asAndroidBitmap().compress(Bitmap.CompressFormat.PNG, 100, outputStream)
+    this.asAndroidBitmap().compress(PNG, 100, outputStream)
     outputStream.flush()
     outputStream.close()
