--- core/common/src/iosMain/kotlin/io/github/droidkaigi/confsched/AppContext.ios.kt
@@ -1,4 +1,0 @@
-package io.github.droidkaigi.confsched
-
-@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
-actual object AppContext
