--- core/ui/src/commonMain/kotlin/io/github/droidkaigi/confsched/ui/SaveToDisk.kt
@@ -1,5 +1,0 @@
-package io.github.droidkaigi.confsched.ui
-
-import androidx.compose.ui.graphics.ImageBitmap
-
-expect fun ImageBitmap.saveToDisk(): String
