--- core/ui/src/iosMain/kotlin/io/github/droidkaigi/confsched/ui/SaveToDisk.ios.kt
@@ -1,8 +1,0 @@
-package io.github.droidkaigi.confsched.ui
-
-import androidx.compose.ui.graphics.ImageBitmap
-
-actual fun ImageBitmap.saveToDisk(): String {
-    // TODO Implement if necessary for use on iOS.
-    return ""
-}
